# docs/swagger.yaml
basePath: /
definitions:
...
  godevmandb.ConClass:
    properties:
      con_class_id:
        type: integer
      created_on:
        type: string
      descr:
        type: string
      notes:
        type: string
      updated_on:
        type: string
    type: object
...
  godevmandb.Connection:
    properties:
      con_cap_id:
        type: integer
      con_class_id:
        type: integer
      con_id:
        type: integer
      con_prov_id:
        type: integer
      con_type_id:
        type: integer
      created_on:
        type: string
      hint:
        type: string
      in_use:
        type: boolean
      notes:
        type: string
      site_id:
        type: integer
      updated_on:
        type: string
    type: object
...
  godevmandb.CreateConClassParams:
    properties:
      descr:
        type: string
      notes:
        type: string
    type: object
...
  godevmandb.UpdateConClassParams:
    properties:
      con_class_id:
        type: integer
      descr:
        type: string
      notes:
        type: string
    type: object
...
  handlers.CountResponse:
    properties:
      count:
        type: integer
    type: object
  handlers.StatusResponse:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
...
info:
  contact:
    email: marko@aretaja.org
    name: API Support
  description: goDevmans API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: goDevmanAPI
  version: v0.0.1-devel.1
paths:
...
  /connections/classes:
    get:
      description: List connection classes info
      operationId: list-con_classes
      parameters:
      - description: url encoded SQL 'ILIKE' operator pattern
        in: query
        name: descr_f
        type: string
      - description: 'min: 1; max: 1000; default: 100'
        in: query
        name: limit
        type: integer
      - description: 'default: 0'
        in: query
        name: offset
        type: integer
      - description: record update time >= (unix timestamp in milliseconds)
        in: query
        name: updated_ge
        type: integer
      - description: record update time <= (unix timestamp in milliseconds)
        in: query
        name: updated_le
        type: integer
      - description: record creation time >= (unix timestamp in milliseconds)
        in: query
        name: created_ge
        type: integer
      - description: record creation time <= (unix timestamp in milliseconds)
        in: query
        name: created_le
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/godevmandb.ConClass'
            type: array
        "404":
          description: Invalid route error
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
        "405":
          description: Invalid method error
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
        "500":
          description: Failde DB transaction
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
      summary: List con_classes
      tags:
      - connections
    post:
      description: Create connection class
      operationId: create-con_class
      parameters:
      - description: JSON object of godevmandb.CreateConClassParams
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/godevmandb.CreateConClassParams'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/godevmandb.ConClass'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
        "404":
          description: Invalid route error
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
        "405":
          description: Invalid method error
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
        "500":
          description: Failde DB transaction
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
      summary: Create con_class
      tags:
      - connections
  /connections/classes/{con_class_id}:
    delete:
      description: Delete connection class
      operationId: delete-con_class
      parameters:
      - description: con_class_id
        in: path
        name: con_class_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid con_class_id
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
        "404":
          description: Invalid route error
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
        "405":
          description: Invalid method error
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
        "500":
          description: Failde DB transaction
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
      summary: Delete con_class
      tags:
      - connections
    get:
      description: Get connection class info
      operationId: get-con_class
      parameters:
      - description: con_class_id
        in: path
        name: con_class_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/godevmandb.ConClass'
        "400":
          description: Invalid con_class_id
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
        "404":
          description: Class not found
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
        "405":
          description: Invalid method error
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
        "500":
          description: Failde DB transaction
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
      summary: Get con_class
      tags:
      - connections
    put:
      description: Update connection class
      operationId: update-con_class
      parameters:
      - description: con_class_id
        in: path
        name: con_class_id
        required: true
        type: string
      - description: JSON object of godevmandb.UpdateConClassParams.<br />Ignored
          fields:<ul><li>con_class_id</li></ul>
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/godevmandb.UpdateConClassParams'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/godevmandb.ConClass'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
        "404":
          description: Invalid route error
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
        "405":
          description: Invalid method error
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
        "500":
          description: Failde DB transaction
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
      summary: Update con_class
      tags:
      - connections
  /connections/classes/{con_class_id}/connections:
    get:
      description: List connection class connections info
      operationId: list-con_class-connections
      parameters:
      - description: con_class_id
        in: path
        name: con_class_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/godevmandb.Connection'
            type: array
        "400":
          description: Invalid con_class_id
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
        "404":
          description: Invalid route error
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
        "405":
          description: Invalid method error
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
        "500":
          description: Failde DB transaction
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
      summary: List con_class connections
      tags:
      - connections
  /connections/classes/count:
    get:
      description: Count number of connection classes
      operationId: count-con_classes
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.CountResponse'
        "404":
          description: Invalid route error
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
        "405":
          description: Invalid method error
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
        "500":
          description: Failde DB transaction
          schema:
            $ref: '#/definitions/handlers.StatusResponse'
      summary: Count con_classes
      tags:
      - connections
...
swagger: "2.0"